feat(backend): remove useless entities
