+++ apps/backend/src/entities/metadata.rs
@@ -1,6 +1,5 @@
 //! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.2
 
-use async_trait::async_trait;
 use chrono::NaiveDate;
 use database::{MetadataLot, MetadataSource};
 use rust_decimal::Decimal;
@@ -46,12 +45,10 @@ pub enum Relation {
     CollectionToEntity,
     #[sea_orm(has_many = "super::metadata_to_genre::Entity")]
     MetadataToGenre,
-    #[sea_orm(has_many = "super::metadata_to_partial_metadata::Entity")]
-    MetadataToPartialMetadata,
+    #[sea_orm(has_many = "super::metadata_to_metadata_group::Entity")]
+    MetadataToMetadataGroup,
     #[sea_orm(has_many = "super::metadata_to_person::Entity")]
     MetadataToPerson,
-    #[sea_orm(has_many = "super::partial_metadata::Entity")]
-    PartialMetadata,
     #[sea_orm(has_many = "super::review::Entity")]
     Review,
     #[sea_orm(has_many = "super::seen::Entity")]
@@ -78,9 +75,9 @@ impl Related<super::metadata_to_genre::Entity> for Entity {
     }
 }
 
-impl Related<super::metadata_to_partial_metadata::Entity> for Entity {
+impl Related<super::metadata_to_metadata_group::Entity> for Entity {
     fn to() -> RelationDef {
-        Relation::MetadataToPartialMetadata.def()
+        Relation::MetadataToMetadataGroup.def()
     }
 }
 
@@ -117,18 +114,17 @@ impl Related<super::genre::Entity> for Entity {
     }
 }
 
-impl Related<super::partial_metadata::Entity> for Entity {
+impl Related<super::metadata_group::Entity> for Entity {
     fn to() -> RelationDef {
-        super::metadata_to_partial_metadata::Relation::PartialMetadata.def()
+        super::metadata_to_metadata_group::Relation::MetadataGroup.def()
     }
     fn via() -> Option<RelationDef> {
         Some(
-            super::metadata_to_partial_metadata::Relation::Metadata
+            super::metadata_to_metadata_group::Relation::Metadata
                 .def()
                 .rev(),
         )
     }
 }
 
-#[async_trait]
 impl ActiveModelBehavior for ActiveModel {}
