+++ apps/backend/src/entities/metadata_to_metadata_group.rs
@@ -1,17 +1,16 @@
-//! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.2
+//! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.6
 
-use database::MetadataToPartialMetadataRelation;
 use sea_orm::entity::prelude::*;
 use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq, Serialize, Deserialize)]
-#[sea_orm(table_name = "metadata_to_partial_metadata")]
+#[sea_orm(table_name = "metadata_to_metadata_group")]
 pub struct Model {
+    pub part: i32,
+    #[sea_orm(primary_key, auto_increment = false)]
+    pub metadata_group_id: i32,
     #[sea_orm(primary_key, auto_increment = false)]
     pub metadata_id: i32,
-    #[sea_orm(primary_key, auto_increment = false)]
-    pub partial_metadata_id: i32,
-    pub relation: MetadataToPartialMetadataRelation,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
@@ -25,13 +24,13 @@ pub enum Relation {
     )]
     Metadata,
     #[sea_orm(
-        belongs_to = "super::partial_metadata::Entity",
-        from = "Column::PartialMetadataId",
-        to = "super::partial_metadata::Column::Id",
+        belongs_to = "super::metadata_group::Entity",
+        from = "Column::MetadataGroupId",
+        to = "super::metadata_group::Column::Id",
         on_update = "Cascade",
         on_delete = "Cascade"
     )]
-    PartialMetadata,
+    MetadataGroup,
 }
 
 impl Related<super::metadata::Entity> for Entity {
@@ -40,9 +39,9 @@ impl Related<super::metadata::Entity> for Entity {
     }
 }
 
-impl Related<super::partial_metadata::Entity> for Entity {
+impl Related<super::metadata_group::Entity> for Entity {
     fn to() -> RelationDef {
-        Relation::PartialMetadata.def()
+        Relation::MetadataGroup.def()
     }
 }
