+++ apps/backend/src/entities/mod.rs
@@ -1,4 +1,4 @@
-//! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.3
+//! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.6
 
 pub mod prelude;
 
@@ -11,12 +11,10 @@ pub mod import_report;
 pub mod metadata;
 pub mod metadata_group;
 pub mod metadata_to_genre;
-pub mod metadata_to_partial_metadata;
+pub mod metadata_to_metadata;
+pub mod metadata_to_metadata_group;
 pub mod metadata_to_person;
-pub mod partial_metadata;
-pub mod partial_metadata_to_metadata_group;
 pub mod person;
-pub mod person_to_partial_metadata;
 pub mod review;
 pub mod seen;
 pub mod user;
