+++ apps/backend/src/entities/prelude.rs
@@ -1,4 +1,4 @@
-//! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.3
+//! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.6
 
 pub use super::calendar_event::Entity as CalendarEvent;
 pub use super::collection::Entity as Collection;
@@ -9,12 +9,10 @@ pub use super::import_report::Entity as ImportReport;
 pub use super::metadata::Entity as Metadata;
 pub use super::metadata_group::Entity as MetadataGroup;
 pub use super::metadata_to_genre::Entity as MetadataToGenre;
-pub use super::metadata_to_partial_metadata::Entity as MetadataToPartialMetadata;
+pub use super::metadata_to_metadata::Entity as MetadataToMetadata;
+pub use super::metadata_to_metadata_group::Entity as MetadataToMetadataGroup;
 pub use super::metadata_to_person::Entity as MetadataToPerson;
-pub use super::partial_metadata::Entity as PartialMetadata;
-pub use super::partial_metadata_to_metadata_group::Entity as PartialMetadataToMetadataGroup;
 pub use super::person::Entity as Person;
-pub use super::person_to_partial_metadata::Entity as PersonToPartialMetadata;
 pub use super::review::Entity as Review;
 pub use super::seen::Entity as Seen;
 pub use super::user::Entity as User;
